R6 Function class.

Class representing a function.

Super class

PaRe::Code -> Function

Methods

Public methods

Function$new()

Function$getFunction()

Function$clone()

Inherited methods

PaRe::Code$getLines()

PaRe::Code$getNLines()

PaRe::Code$getName()

PaRe::Code$print()

Method new()

Initializer for Function object.

Usage

Function$new(name, lineStart, lineEnd, lines)

Arguments

name

<character> Name of Function.

lineStart

<numeric> Line number where function starts in File.

lineEnd

<numeric> Line number where function ends in File.

lines

<c> of type <[base]character> Lines of just the function in File.

Returns

`invisible(self)`

Method getFunction()

Get method to get defined functions in a File object.

Usage

Function$getFunction()

Returns

<data.frame> containing the function name, line start, line end, number of arguments and cyclomatic complexity.

Method clone()

The objects of this class are cloneable with this method.

Usage

Function$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.