whiteList

data.frame containing links to csv-files which should be used to fetch white-listed dependencies.

whiteList

Format

An object of class tbl_df (inherits from tbl, data.frame) with 3 rows and 4 columns.

Details

By default three csv's are listed:

darwin

hades

tidyverse

The data.frame is locally fetched under: system.file(package = "PaRe", "whiteList.csv")

Manual insertions into this data.frame can be made, or the data.frame can be overwritten entirely.

The data.frame itself has the following structure:

column	data type	description

source	character	name of the source

link	character	link or path to the csv-file

package	character	columnname of the package name column in the csv-file being linked to

version	character	columnname of the version column in the csv-file being linked to

The csv-files that are being pointed to should have the following structure:

Examples

if (interactive()) {
  # Dropping tidyverse
  whiteList <- whiteList %>%
    dplyr::filter(source != "tidyverse")

  # getDefaultPermittedPackages will now only use darwin and hades
  getDefaultPermittedPackages()
}